+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+- fix: Set LogLevel in startWithConfigureOptions #613
+
 ## 5.1.10-beta.0
 
 - fix: Abnormal sessions #607